tenzinagent/requirements.txt
# Core dependencies
anthropic>=0.3.0
openai>=1.0.0
langchain>=0.0.300
pydantic>=2.0.0
fastapi>=0.100.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Testing and development
pytest>=7.0.0
black>=23.0.0
bandit>=1.7.0

# Deployment
docker>=6.1.0
python-dotenv>=1.0.0

# ML/AI
torch>=2.0.0  # Changed from pytorch to torch
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0  # CPU version for compatibility

# API and Web
uvicorn>=0.15.0
httpx>=0.24.0
requests>=2.31.0
aiohttp>=3.8.0

# Database and caching
redis>=4.5.0
psycopg2-binary>=2.9.0
asyncpg>=0.27.0

# Monitoring
prometheus-client>=0.16.0
elasticsearch>=8.0.0